fully replaces old PreviewTable with new virtualized version
