--- src/lib/components/table/Table.svelte
@@ -1,3 +1,0 @@
-<table on:mouseleave class="relative border-collapse bg-white">
-  <slot />
-</table>
