+++ src/lib/components/table/TableCells.svelte
@@ -1,0 +1,32 @@
+<script lang="ts">
+  import type { ProfileColumn } from "$lib/types";
+  import Cell from "./Cell.svelte";
+  import Row from "./Row.svelte";
+  export let virtualColumnItems;
+  export let virtualRowItems;
+  export let rows;
+  export let columns: ProfileColumn[];
+  export let scrolling = false;
+  export let activeIndex: number;
+</script>
+
+{#each virtualColumnItems as column (column.key)}
+  <Row>
+    {#each virtualRowItems as row (`${row.key}-${column.key}`)}
+      {@const value = rows[row.index][columns[column.index]?.name]}
+      {@const type = columns[column.index]?.type}
+      {@const rowActive = activeIndex === row?.index}
+      {@const suppressTooltip = scrolling}
+
+      <Cell
+        {suppressTooltip}
+        {rowActive}
+        {value}
+        {row}
+        {column}
+        {type}
+        on:inspect
+      />
+    {/each}
+  </Row>
+{/each}
