+++ src/lib/components/table/config.ts
@@ -1,0 +1,5 @@
+export const config = {
+  columnWidth: 200,
+  rowHeight: 36,
+  indexWidth: 60,
+};
