+++ src/lib/components/table/index.ts
@@ -1,6 +1,1 @@
-import Table from "./Table.svelte";
-import TableRow from "./TableRow.svelte";
-import TableHeader from "./TableHeader.svelte";
-import TableCell from "./TableCell.svelte";
-
-export { Table, TableRow, TableHeader, TableCell };
+export { default as PreviewTable2 } from "./PreviewTable2.svelte";
